requirements.txt
pandas==2.1.4
numpy==1.24.3
sentence-transformers==2.2.2
faiss-cpu==1.7.4
streamlit==1.29.0
PyPDF2==3.0.1
reportlab==4.0.7
groq==0.4.1
python-dotenv==1.0.0
